Add files via upload
This is a code showing how to structure the LLM output to a format I want. Here the output is JSON format and code shows asking LLM to generate the required output requires to show an example to LLM and also using few shot prompt for more accuracy. LLM model strength does not really help in structuring the desired output